Emit subscription when the campaign is updated
