+++ server/routes/graphql/resolvers/subscription.resolver.js
@@ -2,7 +2,7 @@ import config from '../../../config'
 
 export const redeemedCoupon = (params) => {
   return {
-    subscribe: (arg1, arg2, connectionParams) => {
+    subscribe: (arg1, variables, connectionParams) => {
       const hunterId = getUserId(connectionParams);
       if (hunterId) {
         return params.pubsub.asyncIterator(`${config.subscriptionsTopics.REDEEMED_COUPON_TOPIC}-${hunterId}`)
@@ -15,7 +15,7 @@ export const redeemedCoupon = (params) => {
 
 export const huntedCoupon = (params) => {
   return {
-    subscribe: (arg1, arg2, connectionParams) => {
+    subscribe: (arg1, variables, connectionParams) => {
       const makerId = getUserId(connectionParams);
       if (makerId) {
         return params.pubsub.asyncIterator(`${config.subscriptionsTopics.HUNTED_COUPON_TOPIC}-${makerId}`)
@@ -28,12 +28,26 @@ export const huntedCoupon = (params) => {
 
 export const expiredCampaign = (params) => {
   return {
-    subscribe: (arg1, arg2, connectionParams) => {
+    subscribe: (arg1, variables, connectionParams) => {
       const makerId = getUserId(connectionParams);
       if (makerId) {
         return params.pubsub.asyncIterator(`${config.subscriptionsTopics.EXPIRED_CAMPAIGN_TOPIC}-${makerId}`)
       } else {
         throw Error('There is a problem connecting to "expiredCampaign" subscription. Please check the connection params');
+      }
+    }
+  }
+};
+
+export const updatedCampaign = (params) => {
+  return {
+    subscribe: (arg1, variables, connectionParams) => {
+      const makerId = getUserId(connectionParams);
+      const campaignId = variables && variables.campaignId ? variables.campaignId : '';
+      if (makerId) {
+        return params.pubsub.asyncIterator(`${config.subscriptionsTopics.UPDATED_CAMPAIGN_TOPIC}-${campaignId}`)
+      } else {
+        throw Error('There is a problem connecting to "updatedCampaign" subscription. Please check the connection params');
       }
     }
   }
